refactor(agents): move agents module + internals into src/modules/agents/
