+++ src/modules/types.ts
@@ -1,4 +1,4 @@
 export * from "./app.types.js";
-export * from "./agents.types.js";
+export * from "./agents/agents.types.js";
 export * from "./connectors.types.js";
 export * from "./analytics.types.js";
